improved slack post messages
